message store

+++ app/Http/Controllers/Api/ConsultationsController.php
@@ -33,44 +33,37 @@ class ConsultationsController extends Controller
             ], 404);
         }
 
-        $consultation = new Consultation();
-        $consultant = Consultant::with('user')->where('id', $request->consultantId)->first();
-
         try {
-            $consultation->title = $request->title;
-            $consultation->closed = false;
-            $consultation->user_id = $parented->user->id;
-            $consultation->category_id = $request->categoryId;
 
-            $consultation->save();
+        $consultation = new Consultation();
+        $consultation->title = $request->title;
+        $consultation->closed = false;
+        $consultation->user_id = $parented->user->id;
+        $consultation->specialization_id = $request->specializationId;
 
-            $consultation->users()->attach($consultant->user->id);
-            $consultation->users()->attach($parented->user->id);
+        $consultation->save();
 
-            if ($request->allConsultants) {
-                $consultants = Consultant::all();
-                foreach ($consultants as $consultant) {
-                    $consultation->users()->attach($consultant->user->id);
+        $messageData = [
+            'consultation_id' => $consultation->id,
+            'user_id' => $parented->user->id,
+            'text' => $request->messageText,
+            'readed' => false,
+        ];
 
-                    $consultation->messages()->insert([
-                        'consultation_id' => $consultation->id,
-                        'user_id' => $parented->user->id,
-                        'text' => $request->messageText,
-                        'readed' => false
-                    ]);
-                }
-            }
+        if ($request->allConsultants) {
+            $consultants = Consultant::with('user')->where('specialization_id', $request->specializationId)->get();
+            $consultation->users()->attach($consultants->pluck('user.id')->toArray());
+            $consultation->messages()->insert($messageData);
 
-            $consultation->messages()->insert([
-                'consultation_id' => $consultation->id,
-                'user_id' => $parented->user->id,
-                'text' => $request->messageText,
-                'readed' => false
-            ]);
+        } else {
+            $consultant = Consultant::with('user')->where('id', $request->consultantId)->first();
+            $consultation->users()->attach([$consultant->user->id, $parented->user->id]);
+            $consultation->messages()->insert($messageData);
+        }
 
-            return response()->json([
-                'message' => 'Consultation successfully added'
-            ], 200);
+        return response()->json([
+            'message' => 'Consultation successfully added'
+        ], 200);
 
         } catch (\Exception $e) {
             return response()->json([
@@ -80,66 +73,16 @@ class ConsultationsController extends Controller
         }
     }
 
-    public function show(int $id): JsonResource
-    {
-        return ConsultationResource::collection(Consultation::with('users', 'messages')
-        ->where('user_id', Auth::user()->id)
-        ->where('id', $id)->get());
-    }
-
-    public function update(UpdateConsultationRequest $request, string $id): JsonResponse
-    {
-        $parented = Parented::with('user')->where('user_id', auth()->user()->id)->first();
-        if (!$parented) {
-            return response()->json([
-                'message' => 'Only a parent can request a consultation'
-            ], 404);
-        }
-
-        $consultation = Consultation::find($id);
-        $oldConsultant = Consultant::with('user')->where('id', '!=', $consultation->user_id)->first();
-        $newConsultant = Consultant::with('user')->where('id', $request->consultantId)->first();
-
-        try {
-            $consultation->title = $request->title;
-            $consultation->closed = $request->closed;
-            $consultation->user_id = $parented->user->id;
-            $consultation->category_id = $request->categoryId;
-
-            $consultation->save();
-
-            $consultation->users()->detach($oldConsultant->user->id);
-            $consultation->users()->attach($newConsultant->user->id);
-            $consultation->users()->attach($parented->user->id);
-
-            if ($request->allConsultants) {
-                $consultants = Consultant::all();
-                foreach ($consultants as $consultant) {
-                    $consultation->users()->attach($consultant->user->id);
-
-                    $consultation->messages()->insert([
-                        'consultation_id' => $consultation->id,
-                        'user_id' => $parented->user->id,
-                        'text' => $request->messageText,
-                        'readed' => false
-                    ]);
-                }
-            }
-
-            return response()->json([
-                'message' => 'Consultation successfully udated'
-            ], 200);
-
-        } catch (\Exception $e) {
-            return response()->json([
-                'message' => 'Something went wrong in ConsultantationsController.update'
-            ], 400);
-        }
-    }
-
     public function destroy(string $id): JsonResponse
     {
+        $consultation = Consultation::where('id', $id)->first();
+
         try {
+            if ($consultation && $consultation->messages()->count() > 1) {
+                return response()->json([
+                    'message' => 'You can\'t delete a consultation with messages'
+                ], 400);
+            }
             Consultation::destroy($id);
             return response()->json([
                 'message' => 'Consultation successfully deleted'
